[refactor] 테스트 유틸 패키지 이름 변경 #16
